refactor: move PriceDisplay component to components directory
